requirements.txt
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.2
esprima==4.0.1
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
javalang==0.13.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
numpy==2.3.2
python-dotenv==1.1.1
requests==2.32.4
scikit-learn==1.6.1
scipy==1.16.1
six==1.17.0
threadpoolctl==3.6.0
urllib3==2.5.0
Werkzeug==3.1.3
